fix : 'store' 데이터 분리
